feat: Add RotateOnSelf Motion
