+++ navigation-ms/navigation-luffy/navigation/processing/CMakeLists.txt
@@ -2,6 +2,6 @@ robocin_cpp_library(
   NAME processing
   HDRS navigation_processor.h
   SRCS navigation_processor.cpp
-  DEPS navigation_utils
+  DEPS navigation_entities
        motion_parser
 )
