+++ navigation-ms/navigation-luffy/navigation/processing/entities/CMakeLists.txt
@@ -1,5 +1,5 @@
 robocin_cpp_library(
-  NAME navigation_utils
+  NAME navigation_entities
   HDRS robot_move.h
   SRCS robot_move.cpp
   DEPS common::geometry
