+++ navigation-ms/navigation-luffy/navigation/processing/entities/robot_move.cpp
@@ -1,0 +1,1 @@
+#include <navigation/processing/entities/robot_move.h>
